Evolverbetert

Description

Basic data types and classes.

Documentation

initialize :: [Flag] -> IO (IORef World, Handle)

Uses a list of flags to initialize a IORef to a world and a handle for output. Generates an initial agent, populates the world wit it Sets the PRNG

main :: IO ()

Maybe initializes graphical display dependent on display in Parameters calls initialize starts mainLoop

mainLoop :: IORef World -> Handle -> Time -> IO ()

Recursive function that reads the IORef of the world, changes it with newWorld and writes to world back. Also writes output if outputTime Might display graphical output using mainLoopEvent from MyGraphics

chEnv :: Env -> Rand Env

Changes the Environment dependent on nrEnv It works like a clock

newWorld :: World -> Rand World

Changes the environment dependent on envSwitchProb with chEnv Makes new agents with newAssoc

outputString :: World -> Time -> String

The string of data that outputs every outputStep

newAssoc :: World -> ((Int, Int), Agent) -> Rand ((Int, Int), Agent)

uses reproduceAgent to make a new coordinate-Agent association

reproduceAgent :: World -> (Int, Int) -> Rand Agent

The NSF

options :: [OptDescr Flag]

compilerOpts :: [String] -> IO ([Flag], [String])

data Flag

Constructors

Help

OutputFile String

WorldSeed String

AgentSeed String

Graphics

isHelp :: Flag -> Bool

isOutputFile :: Flag -> Bool

isWorldSeed :: Flag -> Bool

isAgentSeed :: Flag -> Bool

isGraphics :: Flag -> Bool