MyRandom

Documentation

type Rand = State PureMT

data PureMT :: *

runRand :: State PureMT a -> PureMT -> (a, PureMT)

getModifyRand :: Rand PureMT

evalRand :: State PureMT a -> PureMT -> a

getBool :: Rand Bool

getDouble :: Rand Double

getRange :: (Int, Int) -> Rand Int

getMyStdRandom :: (PureMT -> (a, PureMT)) -> IO a

setMyStdGen :: PureMT -> IO ()

getMyStdGen :: IO PureMT

Gets the global random number generator.

randomBool :: PureMT -> (Bool, PureMT)

pureMT :: Integral a => a -> PureMT

Module : System.Random.Mersenne.Pure64 Copyright : Copyright (c) 2008, Don Stewart dons@galois.com License : BSD3 Maintainer : Don Stewart dons@galois.com Stability : experimental Portability: CPP, FFI Tested with: GHC 6.8.3

A purely functional binding 64 bit binding to the classic mersenne twister random number generator. This is more flexible than the impure 'mersenne-random' library, at the cost of being a bit slower. This generator is however, many times faster than System.Random, and yields high quality randoms with a long period.

This generator may be used with System.Random, however, that is likely to be slower than using it directly.

Create a PureMT generator from an Integral seed.

Orphan instances

RandomGen PureMT

Methods

next :: PureMT -> (Int, PureMT)

genRange :: PureMT -> (Int, Int)

split :: PureMT -> (PureMT, PureMT)