World

Documentation

data World

Constructors

World

Fields

agents :: Agents

env :: Env

type Env = Int

type Agents = Array (Int, Int) Agent

data Agent

Constructors

Agent

Fields

genome :: Genome

geneStateTable :: GeneStateTable

NoAgent

type Genome = [Chromosome]

type Chromosome = [Locus]

type GeneStateTable = Map ID GeneState

showGST :: GeneStateTable -> String

data Locus

Constructors

Transposon

CGene Gene

CTfbs Tfbs

parseLoc :: [Char] -> Locus

data Gene

Constructors

Gene

Fields

geneID :: ID

thres :: Thres

genSt :: GeneState

data Tfbs

Constructors

Tfbs

Fields

tfbsID :: ID

wt :: Weight

type ID = Int

type Thres = Int

type GeneState = Int

type Weight = Int

class GeneType a where

Minimal complete definition

iD

Methods

iD :: a -> ID

devAg :: Agent -> Agent

Develops agents according to devTime and updateAgent

sameGST :: Agent -> Agent -> Bool

Do two agents share GST

defaultGst :: GeneStateTable

The default GeneStateTable where All genes have expression 0

locusEffect :: Locus -> GeneStateTable -> Int

Check if locus is effected by expressed transcription factors Return the effect it has. 0 is no effect.

updateAgent :: Agent -> Agent

Updates the genestates in the genome and the genestatetable with updateGenome and gSTFromGenome Kills the agent if it doesn't have all genes (when length gst /= nrGeneTypes)

updateGenome :: GeneStateTable -> Genome -> Genome

Updates every Chromosome in Genome with updateChrom

updateChrom :: Int -> GeneStateTable -> Chromosome -> Chromosome

Updates each gene in Chromosome using updateLoc Initial argument is an accumulator that counts the effects of binding sites.

updateLoc :: Int -> GeneStateTable -> Locus -> (Int, Locus)

if Tfbs, increment accumulater with effect if Gene, change the expression according to accumulater and set accumulater to 0 returns a (accumulator, changed Locus) pair

isGene :: Locus -> Bool

Check whether a locus is a Gene

isTfbs :: Locus -> Bool

Check whether a locus is a Gene

getGene :: Locus -> Maybe Gene

Returns Just Gene if Locus is a Gene else Nothing

getTfbs :: Locus -> Maybe Tfbs

Returns Just Tfbs if Locus is a Tfbs else Nothing

groupGeneTfbs :: [Locus] -> [[Locus]]

Groups genes and preceding transcription factors together e.g. [TF, TF, Gene, TF, Gene, TF] -> [[TF, TF, Gene], [TF, Gene], [TF]]

takeWhileInclusive :: (a -> Bool) -> [a] -> ([a], [a])

takes a predicate and returns a pair of lists with the first ending with the element that satisfies the predicate

gSTFromGenome :: Genome -> GeneStateTable

Generate GST from a genome

reduceToTfbss :: Genome -> [Tfbs]

Reduce a genome to a list of its transcription factor binding sites

reduceToGenes :: Genome -> [Gene]

Reduce a genome to a list of its genes

fitnessAgent :: Env -> Agent -> Double

The fitness of an Agent in an Environment (stub for fitnessGST)

fitnessGST :: Env -> GeneStateTable -> Double

Uses targetGST to check fitness of passed GST

hammDist :: Eq a => [a] -> [a] -> Int

Calculate Hamming distance between two lists. For lists with unequal lengths compares only the initial overlap

hammDistAg :: Env -> Agent -> Int

targetGST :: Env -> GeneStateTable

Generate GeneStateTable based on targetExpression

targetExpression :: Env -> ID -> Int

the targetExpression of a Gene in an Environment example for nrEnv = 4 and nrHouseHold = 4, nrOverlap = 3, nrUnique = 5 EnvGene 0 1 2 3 4 5 6 7 8 9 10 11 0 1 1 1 1 0 1 1 1 0 0 0 1 1 1 1 1 1 1 0 1 0 1 0 0 0 2 1 1 1 1 1 1 0 0 0 1 0 0 3 1 1 1 1 0 1 1 0 0 0 1 0

agent0 :: Agent

Just a standard agent that can be used should not be used though, as the tfbs weights have strange weights and such

randomAgent :: Rand Agent

Generate a random Agent using goodRandomGenome

connected :: Genome -> Bool

Does every gene in genome have an associated transcription factor?

goodRandomGenome :: Rand Genome

Generate a random genome that is connected

randomGenome :: Rand Genome

Just a randomChromosome

randomChromosome :: Rand Chromosome

Generate a random chromosome by shuffling twice as many random Tfbss as random genes. using randomTfbss and randomGenes and shuffle'

randomTfbss :: Rand [Locus]

Generate all possible Tfbss (0..nrGeneTypes) with each random weight

randomGenes :: Rand [Locus]

Generate all possible genes (0..nrGeneTypes) with each random threshold and state 0