World

Documentation

devAg :: Agent -> Agent

Develops agents according to devTime and updateAgent

devAg' :: Agent -> Agent

sameGST :: Agent -> Agent -> Bool

Do two agents share GST

defaultGst :: GST

The default GeneStateTable where All genes have expression 0

setToStart :: Agent -> Agent

updateAgent :: Agent -> Agent

Updates the genestates in the genome and the genestatetable with updateGenome and gSTFromGenome Kills the agent if it doesn't have all genes (when length gst /= nrGeneTypes)

updateGenome :: InferGST gst => gst -> Genome -> Genome

Updates every Chromosome in Genome with updateChrom

(↞) :: (b -> c) -> (a1 -> a -> b) -> a1 -> a -> c

updateChrom :: InferGST gst => gst -> Chromosome -> Chromosome

updateTfbss :: InferGST gst => gst -> Chromosome -> Chromosome

updateGenes :: Chromosome -> Chromosome

updateGene :: Integer -> Gene -> Gene

groupGeneTfbs :: [Locus] -> [[Locus]]

Groups genes and preceding transcription factors together e.g. [TF, TF, Gene, TF, Gene, TF] -> [[TF, TF, Gene], [TF, Gene], [TF]]

takeWhileInclusive :: (a -> Bool) -> [a] -> ([a], [a])

takes a predicate and returns a pair of lists with the first ending with the element that satisfies the predicate

connected :: Genome -> Bool

Answers the question: Does every gene of this genome have at least one associated transcription factor?

randomAgent :: Rand Agent

Generate a random Agent using goodRandomGenome

goodRandomGenome :: Rand Genome

Generate a random genome that is connected loops until it finds one

randomGenome :: Rand Genome

Just a randomChromosome

randomChromosome :: Rand Chromosome

Generate a random chromosome by shuffling twice as many random Tfbss as random genes. using randomTfbss and randomGenes and shuffle'

randomTfbss :: Rand [Locus]

Generate all possible Tfbss (0..nrGeneTypes) with each random weight

randomGenes :: Rand [Locus]

Generate all possible genes (0..nrGeneTypes) with each random threshold and state 0

dead :: Agent -> Bool

living :: Agent -> Bool